feat : 포인트 및 등급 조회 api를 가져오기 위한 인터셉터 추가
