+++ src/main/java/com/nhnacademy/frontserver1/common/config/WebConfig.java
@@ -1,0 +1,20 @@
+package com.nhnacademy.frontserver1.common.config;
+
+import com.nhnacademy.frontserver1.common.interceptor.UserInfoInterceptor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebConfig implements WebMvcConfigurer {
+
+    private final UserInfoInterceptor userInfoInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(userInfoInterceptor)
+            .addPathPatterns("/mypage/**");
+    }
+}
