+++ src/main/java/com/nhnacademy/frontserver1/infrastructure/adaptor/CartAdaptor.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.frontserver1.infrastructure.adaptor;
 
-import com.nhnacademy.frontserver1.presentation.dto.request.order.ReadCartBookResponse;
+import com.nhnacademy.frontserver1.presentation.dto.response.order.ReadCartBookResponse;
 import java.util.List;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
