+++ src/main/java/com/nhnacademy/frontserver1/presentation/controller/OrderController.java
@@ -3,11 +3,10 @@ package com.nhnacademy.frontserver1.presentation.controller;
 import com.nhnacademy.frontserver1.application.service.OrderService;
 import com.nhnacademy.frontserver1.domain.TakeoutType;
 import com.nhnacademy.frontserver1.presentation.dto.request.order.CreateOrderRequest;
-import com.nhnacademy.frontserver1.presentation.dto.request.order.ReadCartBookResponse;
 import com.nhnacademy.frontserver1.presentation.dto.request.order.ReadOrderNoneMemberRequest;
 import com.nhnacademy.frontserver1.presentation.dto.request.order.UpdateOrderRequest;
 import com.nhnacademy.frontserver1.presentation.dto.response.order.CreateOrderResponse;
-import com.nhnacademy.frontserver1.presentation.dto.response.order.ReadOrderDeliveryInfoResponse;
+import com.nhnacademy.frontserver1.presentation.dto.response.order.ReadCartBookResponse;
 import com.nhnacademy.frontserver1.presentation.dto.response.order.ReadOrderDetailResponse;
 import com.nhnacademy.frontserver1.presentation.dto.response.order.ReadOrderStatusResponse;
 import com.nhnacademy.frontserver1.presentation.dto.response.order.ReadOrderUserAddressResponse;
@@ -137,22 +136,6 @@ public class OrderController {
     public ResponseEntity<UpdateOrderResponse> updateOrder(@PathVariable String orderId,
         @RequestBody UpdateOrderRequest request) {
         return ResponseEntity.ok(orderService.updateOrderByOrderId(orderId, request));
-    }
-
-    @GetMapping("/{orderId}")
-    public String getOrder(@PathVariable String orderId, Model model) {
-        ReadOrderDetailResponse response = orderService.getMyOrderByOrderId(orderId);
-        model.addAttribute("order", response);
-
-        return "mypage/mypage-orders-detail";
-    }
-
-    @GetMapping("/{orderId}/delivery")
-    public String getOrderDelivery(@PathVariable String orderId, Model model) {
-        ReadOrderDeliveryInfoResponse orderInfoResponse = orderService.getMyOrderDelivery(orderId);
-        model.addAttribute("orderInfo", orderInfoResponse);
-
-        return "mypage/mypage-delivery-detail";
     }
 
     @GetMapping("/find")
